Extract Topics:Multiple test
